requirements.txt
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
black==24.8.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.4
click==8.1.7
decorator==5.1.1
distlib==0.3.8
Django==5.1.1
django-unicorn==0.61.0
filelock==3.15.4
Flask==3.0.3
h11==0.14.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mypy-extensions==1.0.0
mysqlclient==2.2.6
numpy==1.26.4
orjson==3.10.7
outcome==1.3.0.post0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pycodestyle==2.12.1
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
selenium==4.25.0
shortuuid==1.0.13
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
